Flatten moq-web a bit.
