+++ .editorconfig
@@ -6,7 +6,7 @@ end_of_line = lf
 trim_trailing_whitespace = true
 insert_final_newline = true
 indent_style = tab
-indent_size = 2
+indent_size = 4
 max_line_length = 120
 
 # YAML doesn't support hard tabs 🙃
