--- .vscode/extensions.json
@@ -1,6 +1,0 @@
-{
-	"recommendations": [
-		"EditorConfig.EditorConfig",
-		"rust-lang.rust-analyzer",
-	]
-}
