+++ README.md
@@ -3,41 +3,46 @@
 </p>
 
 Media over QUIC (MoQ) is a live media delivery protocol utilizing QUIC.
-It's a client-server model (not peer-to-peer) that is designed to scale to enormous viewership via clustered relay servers (aka a CDN).
+It's a client-server model that is designed to scale to enormous viewership via clustered relay servers (aka a CDN).
 The application determines the trade-off between latency and quality, potentially on a per viewer-basis.
 
 See [quic.video](https://quic.video) for more information.
-Note: this project uses a forked implementation of the IETF standardization effort.
+Note: this project is a [fork of the IETF draft](https://quic.video/blog/transfork) to speed up development.
 
 The project is split into a few crates:
 
--   [moq-transfork](moq-transfork): The underlying network protocol. It's designed for media-like applications that need real-time and scale.
 -   [moq-relay](moq-relay): A server that forwards content from publishers to any interested subscribers. It can optionally be clustered, allowing N servers to transfer between themselves.
-- [moq-karp](moq-karp): A simple media layer powered by moq-transfork, intended as a HLS/DASH/SDP replacement. It includes a CLI for converting between formats.
+- [moq-web](moq-web): A web client utilizing Rust and WASM. Supports both consuming media (and soon publishing).
+-   [moq-transfork](moq-transfork): The underlying network protocol. It can be used by live applications that need real-time and scale, even if they're not media.
+- [moq-karp](moq-karp): The underlying media protocol powered by moq-transfork. It includes a CLI for importing/exporting to other formats, for example integrating with ffmpeg.
 -   [moq-clock](moq-clock): A dumb clock client/server just to prove MoQ can be used for more than media.
 -   [moq-native](moq-native): Helpers to configure the native MoQ tools.
 
 There are additional components that have been split into other repositories for development reasons:
 
 - [moq-gst](https://github.com/kixelated/moq-gst): A gstreamer plugin for producing Karp broadcasts.
-- [moq-wasm](https://github.com/kixelated/moq-wasm): A web client utilizing Rust and WASM.
-- [moq-js](https://github.com/kixelated/moq-js): A web client utilizing Typescript.
-
-At the moment [moq-js](https://github.com/kixelated/moq-js) contains the only mature player.
-A hosted version is available at [quic.video](https://quic.video) and you can use the `?host=localhost:4443` query parameter to target a local moq-relay.
 
 # Usage
-For quick iteration cycles, use the [dev helper scripts](dev/README.md).
-
-To launch a full cluster, including provisioning certs and deploying root certificates, you can use docker-compose via:
-
+To launch a full cluster with docker-compose, run:
 ```
 make run
 ```
 
-Then, visit https://quic.video/publish/?server=localhost:4443.
+Or for faster iteration cycles, use the dev helper scripts:
+```sh
+# Run as a single (hacky) command:
+./dev/all
 
-# Usage
+# Or individually:
+./dev/relay
+./dev/pub
+./dev/web
+```
+
+Then, visit [https://localhost:8080](localhost:8080) to use the demo client.
+
+
+# Components
 
 ## moq-relay
 
@@ -53,6 +58,33 @@ Notable arguments:
 
 This listens for WebTransport connections on `UDP https://localhost:4443` by default.
 You need a client to connect to that address, to both publish and consume media.
+
+## moq-web
+
+[moq-web](moq-web) is a web client that can consume media (and soon publish).
+It's available [on NPM](https://www.npmjs.com/package/@kixelated/moq) as both a JS library and web component.
+
+For example:
+
+```html
+<script type="module">
+	import '@kixelated/moq'
+</script>
+
+<moq-video src="https://relay.quic.video/demo/bbb"></moq-video>
+```
+
+Because it uses WASM, it's not part of the main workspace and requires some extra steps to build.
+
+```sh
+cd moq-web
+bun i
+bun dev
+```
+
+This will start a development server on `http://localhost:3000`.
+There's two separate compilation steps, the first building with `wasm-pack` and the second bundling with `rspack`.
+See the [moq-web README](moq-web/README.md) for more information.
 
 ## moq-karp
