+++ dev/web
@@ -1,0 +1,11 @@
+#!/usr/bin/env bash
+set -euo pipefail
+
+# Change directory to root folder
+cd "$(dirname "$0")/.."
+
+# Install any new dependencies
+bun i
+
+# Start the development server
+bun dev
